+++ db/crud.py
@@ -1,5 +1,4 @@
 from sqlalchemy.orm import Session
-import db.schemas as schemas, db.models as models
 from fastapi import HTTPException, status
 from fastapi_jwt_auth import AuthJWT
 from fastapi.encoders import jsonable_encoder
@@ -63,80 +62,3 @@ def generate_tokens(id: int, authorize: AuthJWT):
         "message": "Success",
     }
     return jsonable_encoder(response)
-
-
-def signup_caregiver(
-    db: Session, authorize: AuthJWT, caregiver: schemas.SignUpCareGiver
-):
-    # Check for mail if it is stored in the db or not
-    db_email = (
-        db.query(models.CareGiver)
-        .filter(models.CareGiver.email == caregiver.email)
-        .first()
-    )
-
-    # * if the email is already exist raise an HTTPException with status code 400 and a message
-    if db_email is not None:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="This email is already exist",
-        )
-    db_username = (
-        db.query(models.CareGiver)
-        .filter(models.CareGiver.username == caregiver.username)
-        .first()
-    )
-
-    # * if the username is already exist raise an HTTPException with status code 400 and a message
-    if db_username is not None:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="This username is already exist",
-        )
-
-    db_id = (
-        db.query(models.CareGiver).filter(models.CareGiver.id == caregiver.id).first()
-    )
-
-    # * Create a new instance of CareGiver model to store the caregiver in the database
-
-    caregiver.id = int(caregiver.id)
-    caregiver.age = int(caregiver.age)
-    caregiver.password = create_hashed_password(caregiver.password)
-
-    new_caregiver = models.CareGiver(**caregiver.dict())
-
-    # * Add the new caregiver to the database session and commit the changes to the database
-    db.add(new_caregiver)
-    db.commit()
-    db.refresh(new_caregiver)
-
-    return generate_tokens(authorize=authorize, id=new_caregiver.id)
-
-
-def get_caregiver(db: Session, authorize: AuthJWT, caregiver: schemas.Login):
-    db_caregiver = (
-        db.query(models.CareGiver)
-        .filter(caregiver.email == models.CareGiver.email)
-        .first()
-    )
-
-    if db_caregiver and verify_password(caregiver.password, db_caregiver.password):
-        return generate_tokens(authorize=authorize, id=db_caregiver.id)
-    return None
-
-
-def login_caregiver(db: Session, authorize: AuthJWT, caregiver: schemas.Login):
-    db_caregiver = get_caregiver(db=db, authorize=authorize, caregiver=caregiver)
-    if db_caregiver is None:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid email or password"
-        )
-    return db_caregiver
-
-
-def caregiver_info(db: Session, caregiver_id: int):
-    caregiver = (
-        db.query(models.CareGiver).filter(models.CareGiver.id == caregiver_id).first()
-    )
-    return schemas.CareGiverInfo.from_orm(caregiver)
